ONScripter engine Korean support (hangul) is added.
Dodges UTF8 collisions by reading the entire line (checking for Korean)
before deciding to show UTF8 or Hangul

It will read the entire string and break out once non-Korean characters
are displayed. Therefore if the entire string is Korean, it will be
displayed as Korean, otherwise it will display the normal ASCII characters
with the UTF-8 characters.

Notice that if the entire string is UTF8, it will collide with the Korean
character detection and will fail (think the entire UTF8 string is
Korean). There is no other way to fully detect if bytes are Korean or
UTF8.

Also this does not support Korean (Hangul) character set encoded in UTF8,
only EUC-KR character set works right now.