Implemented game tracing and playback
This is temporary and only used to track games so I can fix them.
Trace happens whenever you do not playback; therefore anytime you
play the game, it will trace. It will start a new trace when you
load a game. To start playback, you need the trace.log file and
save#.dat file into the downloads folder and long press the
corresponding game in the main menu and click playback (on debug
builds). Then it will automatically load the game (after 2sec)
and will eventually go to the crash.