Fixes glitch for moving save files to internal storage when no files exist
Fixes use case: user has new game with no save files in it on ext sd
card and tries to play it. Launches dialog to pick how to fix saving
and chooses to copy only the save files into the internal storage.
Since there is no files, it shows up as 0 bytes which causes it not
to copy. This fixes that.