Fixes freezing after not choosing an ext video
When user chooses external video from settings and cancels, it used to
freeze. Now we track when the user uses an external video and when the
popup of choosing which video player occurs, we will figure out when to
create a new gl surface so that all possibilities of video playback would
not freeze after it is finished playing.